Renamed files GOGUISize... to GOSize...
